fix: explicit widening on int-math whose result feeds a long/float (S2184)
20 SonarCloud S2184 sites where an int subtraction/addition is widened
to long or float in the surrounding context. Behaviour was correct for
the actual value ranges (page row counts, sample indices, weights all
well below INT_MAX/2), but the intermediate int result could overflow
on adversarial input. Adding the cast on one operand documents the
intent and rules out the edge case.

Touched:
- AlpEncodingEncoder: {@code Math.min(i * stride, n - 1)} → cast n
- PcoEncodingEncoder: {@code Long.numberOfLeadingZeros(n - 1)} /
  {@code n - 2} + {@code writeBits(weights[i] - 1, ...)} → cast operand
- PcoEncodingDecoder: {@code Math.clamp(nRemaining - deltaOrder, ...)}
  re-routed through long-overload, cast back to int for the slot
- PcoBinOptimizer: {@code float = int + int + int} cast first operand
- BitpackedEncodingDecoder: long[] assignment from
  {@code FL_ORDER[o] * 16 + s * 128} → cast operands to long
- Dict/Fsst/VarBin encoders: {@code setAtIndex(.., i + 1, ..)} → cast i

+++ reader/src/main/java/io/github/dfa1/vortex/reader/decode/BitpackedEncodingDecoder.java
@@ -128,7 +128,7 @@ public final class BitpackedEncodingDecoder implements EncodingDecoder {
             nextWordByteBase[row] = rem > 0 ? (long) lanes * nextWord : 0L;
             int o = row / 8;
             int s = row % 8;
-            outRowByteOff[row] = FL_ORDER[o] * 16 + s * 128;
+            outRowByteOff[row] = (long) FL_ORDER[o] * 16 + (long) s * 128;
         }
 
         long blockByteOff = 0L;
